refactor: 调整 core.ts 导出的模块内容以及 index.ts 的模块导入
